Overhaul `block-editor-classes.js` Meant to:
- Fix reliance on `wp.blocks.registerBlockStyle()` to trigger the listener
- Remove flash of unstyled content prior to classes being added

The flash of unstyled content is still present in the legacy version of the block editor that doesn't load the editor itself in an `<iframe>` (as triggered by older blocks). From my testing, I'm not convinced there's a reliable way to fix it in this mode, but I'm happy that under normal conditions, this is now resolved.